Fix auto start sound feedback not playing in some cases
When voice announcements for auto start where enabled but no TTS engine available, the app should play sounds during the last 10s of the countdown. However, this didn't happen, since the sound feedback controller did only check for voice announcements settings and not for the availability of a TTS engine in general.