edit start/end: keep showing the old track (in grey) to allow selection
The code was removing the original workout before displaying the new
workout but this gave two probems. First it was not longer possible
to select the second point using the map as only the original layer
has a selection listener. The second problem this resolves is that
it now remains possible to move the start or end of the workout
back and forth.