Issue #232: introduce free moving in GPS recording map
A threshold will be checked to determine how far the user dragged the
map from center. If the threshold exceeds, auto-focus is disabled.
To enable auto focus, either move the center of the map near to latest
GPS position or click on the center of the view.

The threshold is currently a fixed value in meters but should be changed
to be dependent on zoom level